backend/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-multipart>=0.0.6
email-validator>=2.0.0

# Database
sqlalchemy>=2.0.22
alembic>=1.12.0
psycopg2-binary>=2.9.9
asyncpg>=0.28.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pyotp>=2.9.0
qrcode>=7.4.2
pillow>=10.0.1

# Caching
redis>=5.0.1
valkey-py>=1.0.0

# HTTP Client
httpx>=0.25.0
aiohttp>=3.8.6

# File Processing
python-magic>=0.4.27
rarfile>=4.0
zipfile36>=0.1.3
py7zr>=0.20.5

# Image Processing
pillow>=10.0.1

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tenacity>=8.2.3
structlog>=23.2.0
rich>=13.6.0

# Testing
pytest>=7.4.2
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.25.0

# Development
black>=23.9.1
isort>=5.12.0
flake8>=6.1.0
mypy>=1.6.1
